+++ packages/core/lib/src/core_widget_factory.dart
@@ -1149,6 +1149,11 @@ class WidgetFactory extends WidgetFactoryResetter with AnchorWidgetFactory {
 
     final baseUrl = _widget?.baseUrl;
     if (baseUrl == null) {
+      if (uri.hasAuthority) {
+        // special handling for protocol relative URL without base: assume HTTPS
+        return Uri(scheme: 'https').resolveUri(uri).toString();
+      }
+
       return null;
     }