+++ packages/fwfh_chewie/CHANGELOG.md
@@ -15,6 +15,10 @@
 - Add support for flutter_widget_from_html_core@0.11.0 (#899)
 - Requires Flutter 3.7
 
+## 0.7.1+4
+
+- Add screenshots, funding and topics to pubspec files (#1010)
+
 ## 0.7.1+2
 
 - Add support for Flutter 3.10 (#889)